+++ Fun.Blazor.Docs.Wasm/Routes.fs
@@ -1,4 +1,5 @@
-﻿[<AutoOpen>]
+﻿// hot-reload
+[<AutoOpen>]
 module Fun.Blazor.Docs.Wasm.Routes
 
 open System